fix: add common test helper
