+++ packages/amplify-e2e-tests/src/__tests__/api_6c.test.ts
@@ -4,13 +4,12 @@ import {
   addApiWithoutSchema,
   amplifyPush,
   deleteProjectDir,
-  putItemInTable,
-  scanTable,
   rebuildApi,
   getProjectMeta,
   updateApiSchema,
   amplifyDeleteWithLongerTimeout,
 } from '@aws-amplify/amplify-e2e-core';
+import { testTableAfterRebuildApi, testTableBeforeRebuildApi } from '../rebuild-test-helpers';
 
 const projName = 'apitest';
 
@@ -24,7 +23,7 @@ afterEach(async () => {
 });
 
 describe('amplify rebuild api', () => {
-  it('recreates tables and opensearch service for searchable models', async () => {
+  it('recreates tables for searchable models', async () => {
     await initJSProjectWithProfile(projRoot, { name: projName });
     await addApiWithoutSchema(projRoot, { transformerVersion: 2 });
     await updateApiSchema(projRoot, projName, 'searchable_model_v2.graphql');
@@ -34,12 +33,8 @@ describe('amplify rebuild api', () => {
     const region = projMeta?.providers?.awscloudformation?.Region;
     expect(apiId).toBeDefined();
     expect(region).toBeDefined();
-    const tableName = `Todo-${apiId}-integtest`;
-    await putItemInTable(tableName, region, { id: 'this is a test value' });
-    const scanResultBefore = await scanTable(tableName, region);
-    expect(scanResultBefore.Items.length).toBe(1);
+    await testTableBeforeRebuildApi(apiId, region, 'Todo');
     await rebuildApi(projRoot, projName);
-    const scanResultAfter = await scanTable(tableName, region);
-    expect(scanResultAfter.Items.length).toBe(0);
+    await testTableAfterRebuildApi(apiId, region, 'Todo');
   });
 });
